Release 1.25.0

+++ src/promotion/promotion.service.ts
@@ -195,6 +195,8 @@ export class PromotionService {
           'tier.description',
           'tier.tier_postter',
           'tier.detail',
+          'promotion.promo_id',
+          'promotion.promo_name',
 
           'condition.cond_id',
 
@@ -264,6 +266,21 @@ export class PromotionService {
             status: true,
             start_date: LessThanOrEqual(startOfDay),
             end_date: MoreThanOrEqual(endOfDay),
+          },
+        },
+        relations: {
+          promotion: true,
+        },
+        select: {
+          tier_id: true,
+          tier_name: true,
+          tier_postter: true,
+          min_amount: true,
+          description: true,
+          all_products: true,
+          promotion: {
+            promo_id: true,
+            promo_name: true,
           },
         },
       });
@@ -514,6 +531,23 @@ export class PromotionService {
 
   async createCondition(data: { tier_id: number; product_gcode: string }) {
     try {
+      const tier = await this.promotionTierRepo.findOne({
+        where: { tier_id: data.tier_id },
+        relations: ['promotion'],
+      });
+
+      const findTierisProduct = await this.promotionTierRepo
+        .createQueryBuilder('tier')
+        .leftJoin('tier.promotion', 'promotion')
+        .where('tier.all_products = true')
+        .andWhere('tier.promotion.promo_id = :promo_id', {
+          promo_id: tier?.promotion?.promo_id,
+        })
+        .getMany();
+
+      if (findTierisProduct.length > 0)
+        return 'Cannot set all products for this tier because there are other tiers with the same minimum amount that are not active';
+
       const newCondition = this.promotionConditionRepo.create({
         tier: { tier_id: data.tier_id },
         product: { pro_code: data.product_gcode },
@@ -751,6 +785,23 @@ export class PromotionService {
 
   async setAllProducts(tier_id: number, status: boolean) {
     try {
+      const tier = await this.promotionTierRepo.findOne({
+        where: { tier_id },
+        relations: ['promotion'],
+      });
+
+      const findTierisNotProduct = await this.promotionTierRepo
+        .createQueryBuilder('tier')
+        .leftJoin('tier.promotion', 'promotion')
+        .where('tier.all_products = false')
+        .andWhere('tier.promotion.promo_id = :promo_id', {
+          promo_id: tier?.promotion?.promo_id,
+        })
+        .getMany();
+
+      if (findTierisNotProduct.length > 1)
+        return 'Cannot set all products for this tier because there are other tiers with the same minimum amount that are not active';
+
       if (status === true) {
         await this.promotionConditionRepo.delete({ tier: { tier_id } });
         await this.promotionTierRepo.update(tier_id, {
@@ -797,6 +848,8 @@ export class PromotionService {
         .andWhere('promotion.end_date >= :startOfDay', { startOfDay })
         .select([
           'condition.cond_id',
+          'promotion.promo_id',
+          'promotion.promo_name',
           'tier.tier_id',
           'tier.tier_name',
           'tier.min_amount',
@@ -861,6 +914,21 @@ export class PromotionService {
             status: true,
             start_date: LessThanOrEqual(startOfDay),
             end_date: MoreThanOrEqual(endOfDay),
+          },
+        },
+        relations: {
+          promotion: true,
+        },
+        select: {
+          tier_id: true,
+          tier_name: true,
+          tier_postter: true,
+          min_amount: true,
+          description: true,
+          all_products: true,
+          promotion: {
+            promo_id: true,
+            promo_name: true,
           },
         },
       });
